Code optimization

+++ Source/GAME/drawing_func.h
@@ -17,10 +17,10 @@
 
 void drawScore();
 void drawTimeLeft();
-void drawLives(int initialX, int initialY);
-void drawPowerPill(int xCoord, int yCoord);
-void drawBoard(int drawAllBoardFlag);
-void drawPacmanMove(int newY, int newX, int prevY, int prevX);
+void drawLives(uint16_t initialX, uint16_t initialY);
+void drawPowerPill(uint16_t xCoord, uint16_t yCoord);
+void drawBoard(uint16_t drawAllBoardFlag);
+void drawPacmanMove(uint16_t newY, uint16_t newX, uint16_t prevY, uint16_t prevX);
 void drawGameOverScreen();
 void clearGameOverScreen();
 void drawGameEndedScreen();
